+++ src/app/components/views/SubmitRoutes.js
@@ -3,14 +3,14 @@ import React from "react"
 import { Switch, Route } from "react-router-dom"
 
 // components
-import Intro from "./Intro"
+import Submit from "./Submit"
 import Composer from "./Composer"
 
 // render
-export const Submit = props => {
+export default props => {
 	return (
 		<Switch>
-			<Route exact path="/submit" component={ Intro } />
+			<Route exact path="/submit" component={ Submit } />
 			<Route exact path="/submit/compose" component={ Composer } />
 		</Switch>
 	)
