+++ src/app/containers/Nav.js
@@ -16,20 +16,18 @@ const Nav = props => {
     </Wrapper>
   )
 }
-class NavWrapper extends React.Component {
-  render() {
-		return(
-      <div>
-        { this.props.nav.location.top
-          && <Nav view={ this.props.nav.view } draftStatus={ this.props.composer.draftStatus } />
-        }
-          { this.props.children }
-        { this.props.nav.location.bottom
-          && <Nav view={ this.props.nav.view } draftStatus={ this.props.composer.draftStatus } />
-        }
-      </div>
-    )
-	}
+const NavWrapper = props => {
+	return(
+    <div>
+      { props.nav.location.top
+        && <Nav view={ props.nav.view } draftStatus={ props.composer.draftStatus } />
+      }
+        { props.children }
+      { props.nav.location.bottom
+        && <Nav view={ props.nav.view } draftStatus={ props.composer.draftStatus } />
+      }
+    </div>
+  )
 }
 
 // connet with redux
@@ -39,4 +37,4 @@ const mapStateToProps = state => {
     composer: state.composer,
 	}
 }
-export default connect(mapStateToProps, null, null, {pure: false})(NavWrapper)
+export default connect(mapStateToProps)(NavWrapper)
